Changed required API level to 16 (4.1.1 - Jelly Bean) to take full access on the latest features.
The first Jelly Bean Feature is the availability of the SYSTEM_UI_FLAG_FULLSCREEN flag.
With this flag it is now possible to hide any navigation buttons. This will be useful for providing an AlarmApp that can not be closed, killed or dismiss by accident.
A first look on how this will work is also provided by this commitment.

ToDo:

- Catching system layout changes that are bringing back the navigation buttons -> see http://developer.android.com/reference/android/view/View.OnSystemUiVisibilityChangeListener.html

- Providing backward compatibility of our AlarmApp for older Android versions that a) are widely used at the market and b) that has no navigation buttons drawn directly onto the main display (different from the Galaxy Nexus).

- Deactivating other (soft-) keys on an Android device to prevent closing or dismissing the AlarmApp Activity.

Note: Disallowing to close, kill or dismiss an Android application is not how an app should work!!! But for outer intents this is exactly what we need and what makes the AlarmApp more reliable for a help seeking person.